+++ src/datalevin/client.clj
@@ -227,6 +227,8 @@
       (u/raise "Unable to copy in:" e
                {:req req :count (count data)}))))
 
+(declare open-database)
+
 (deftype ^:no-doc Client [username password host port pool-size time-out
                           ^:volatile-mutable ^UUID id
                           ^:volatile-mutable ^ConnectionPool pool]
@@ -239,7 +241,7 @@
               res  (when-let [{:keys [type] :as result}
                               (try
                                 (send-n-receive conn req)
-                                (catch Exception e
+                                (catch Exception _
                                   (close conn)
                                   nil)
                                 (finally (release-connection pool conn)))]
@@ -248,6 +250,10 @@
                        :copy-out-response (copy-out conn req)
                        :command-complete  result
                        :error-response    result
+                       :reopen
+                       (let [{:keys [db-name db-type]} result]
+                         (vreset! success? false)
+                         (open-database client db-name db-type))
                        :reconnect
                        (let [client-id
                              (authenticate host port username password)]