+++ test/datalevin/core_test.clj
@@ -466,31 +466,36 @@
     (u/delete-files dir)))
 
 (deftest map-resize-clear-test
-  (let [dir  (u/tmp-dir (str "clear-test-" (UUID/randomUUID)))
-        conn (sut/create-conn
-               dir {:buggy/key {:db/valueType :db.type/string
-                                :db/unique    :db.unique/identity}}
-               {:kv-opts {:mapsize 1}})]
-    (sut/transact! conn (for [i (range 100000)]
-                          {:buggy/key (format "%20d" i)
-                           :buggy/val (format "bubba-%d" i)}))
-    (is (= 200000 (count (sut/datoms (sut/db conn) :eav))))
-    (sut/transact! conn [{:buggy/key (format "%20d" 100000)
-                          :buggy/val (format "bubba-%d" 99)}])
-    (sut/close conn)
-    (is (sut/closed? conn))
+  (let [dir (u/tmp-dir (str "clear-test-" (UUID/randomUUID)))]
+    (dotimes [_ 10]
+      (let [conn (sut/create-conn
+                   dir {:buggy/key {:db/valueType :db.type/string
+                                    :db/unique    :db.unique/identity}}
+                   {:kv-opts {:mapsize 1}})]
+        (sut/transact! conn (for [i (range 100000)]
+                              {:buggy/key  (format "%20d" i)
+                               :buggy/val  (format "bubba-%d" i)
+                               :buggy/time (System/currentTimeMillis)}))
+        (is (= 300000 (count (sut/datoms (sut/db conn) :eav))))
+        (sut/close conn)
+        (is (sut/closed? conn))
 
-    (let [conn1 (sut/create-conn dir)]
-      (sut/clear conn1)
-      (is (sut/closed? conn1)))
+        (let [conn1 (sut/create-conn dir)]
+          (sut/clear conn1)
+          (is (sut/closed? conn1)))
 
-    (let [conn2 (sut/create-conn dir)]
-      (is (= 0 (count (sut/datoms (sut/db conn2) :eav))))
-      (sut/transact! conn2 [{:buggy/key (format "%20d" 100001)
-                             :buggy/val (format "bubba-%d" 100)}])
+        (let [conn2 (sut/create-conn dir)]
+          (is (= 0 (count (sut/datoms (sut/db conn2) :eav))))
+          (sut/transact! conn2 [{:buggy/key  (format "%20d" 100001)
+                                 :buggy/val  (format "bubba-%d" 100)
+                                 :buggy/time (System/currentTimeMillis)}])
 
-      (is (= 2 (count (sut/datoms (sut/db conn2) :eav))))
-      (sut/close conn2))
+          (is (= 3 (count (sut/datoms (sut/db conn2) :eav))))
+          (sut/close conn2))
+
+        (let [conn1 (sut/create-conn dir)]
+          (sut/clear conn1)
+          (is (sut/closed? conn1)))))
     (u/delete-files dir)))
 
 (deftest with-txn-map-resize-test